* Move the csi/pkg to top-level pkg. * Update the golang-1.19 to gcr.io/distroless/base:debug as csidriver's base image. * Improve the csidriver example readme.
