+++ k8s/config/crd/kustomization.yaml
@@ -35,7 +35,6 @@ patchesStrategicMerge:
 - patches/cainjection_in_backups.yaml
 - patches/cainjection_in_recovers.yaml
 - patches/cainjection_in_csidrivers.yaml
-#- patches/cainjection_in_csidrivers.yaml
 # +kubebuilder:scaffold:crdkustomizecainjectionpatch
 
 # the following config is for teaching kustomize how to do kustomization for CRDs.
