+++ k8s/pkg/templates/csidriver/deployment.yaml
@@ -26,7 +26,9 @@ spec:
           - --endpoint=$(CSI_ENDPOINT)
           - --nodeid=$(KUBE_NODE_NAME)
           - --state-file-path=/csi/state
+          {{ if .Spec.EnableDebugLog }}
           - --verbose
+          {{ end }}
           env:
             - name: CSI_ENDPOINT
               value: unix:///csi/csi.sock
